Separating Shop and ShopItem classes
